Add AbstractSynchronousGeneratorBuilder and getUPuVarName to transformer model for private synchronous model
