+++ dynawo-simulation/src/main/java/com/powsybl/dynawo/models/generators/SynchronousGeneratorBuilder.java
@@ -7,7 +7,6 @@
  */
 package com.powsybl.dynawo.models.generators;
 
-import com.powsybl.commons.PowsyblException;
 import com.powsybl.commons.report.ReportNode;
 import com.powsybl.dynawo.builders.*;
 import com.powsybl.dynawo.commons.DynawoVersion;
@@ -18,7 +17,7 @@ import java.util.Collection;
 /**
  * @author Laurent Issertial {@literal <laurent.issertial at rte-france.com>}
  */
-public class SynchronousGeneratorBuilder extends AbstractGeneratorBuilder<SynchronousGeneratorBuilder> {
+public class SynchronousGeneratorBuilder extends AbstractSynchronousGeneratorBuilder<SynchronousGeneratorBuilder> {
 
     public static final String CATEGORY = "SYNCHRONOUS_GENERATOR";
     private static final ModelConfigs MODEL_CONFIGS = ModelConfigsHandler.getInstance().getModelConfigs(CATEGORY);
@@ -61,31 +60,6 @@ public class SynchronousGeneratorBuilder extends AbstractGeneratorBuilder<Synchr
 
     protected SynchronousGeneratorBuilder(Network network, ModelConfig modelConfig, ReportNode reportNode) {
         super(network, modelConfig, reportNode);
-    }
-
-    protected EnumGeneratorComponent getGeneratorComponent() {
-        boolean aux = modelConfig.hasAuxiliary();
-        boolean transformer = modelConfig.hasTransformer();
-        if (aux && transformer) {
-            return EnumGeneratorComponent.AUXILIARY_TRANSFORMER;
-        } else if (transformer) {
-            return EnumGeneratorComponent.TRANSFORMER;
-        } else if (aux) {
-            throw new PowsyblException("Generator component auxiliary without transformer is not supported");
-        }
-        return EnumGeneratorComponent.NONE;
-    }
-
-    @Override
-    public SynchronousGenerator build() {
-        if (isInstantiable()) {
-            if (modelConfig.isControllable()) {
-                return new SynchronousGeneratorControllable(getEquipment(), parameterSetId, modelConfig, getGeneratorComponent());
-            } else {
-                return new SynchronousGenerator(getEquipment(), parameterSetId, modelConfig, getGeneratorComponent());
-            }
-        }
-        return null;
     }
 
     @Override
